requirements.txt
playwright
openai
pydantic
python-dotenv
pillow
